serverless/azure-functions/requirements.txt
# Azure Functions Python Requirements

azure-functions>=1.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
